// src/infrastructure/grpc/generated/course/types/stats.ts
export interface EnrollmentTrend {
  month: number;
  enrollments: number;
}

export interface GetInstructorCourseEnrollmentSummeryRequest {
  instructorId: string;
  courseId: string;
}

export interface InstructorCourseEnrollmentTrend {
  trend: EnrollmentTrend[];
}
// ...
export interface GetInstructorCourseEnrollmentTrendRequest {
  courseId: string;
  instructorId: string;
  from: string;
  to: string;
}
// ...
export const EnrollmentTrend: MessageFns<EnrollmentTrend> = {
  encode(
    message: EnrollmentTrend,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.month !== 0) {
      writer.uint32(8).int32(message.month);
    }
    if (message.enrollments !== 0) {
      writer.uint32(16).int32(message.enrollments);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): EnrollmentTrend {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEnrollmentTrend();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.month = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.enrollments = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): EnrollmentTrend {
    return {
      month: isSet(object.month) ? globalThis.Number(object.month) : 0,
      enrollments: isSet(object.enrollments)
        ? globalThis.Number(object.enrollments)
        : 0,
    };
  },

  toJSON(message: EnrollmentTrend): unknown {
    const obj: any = {};
    if (message.month !== 0) {
      obj.month = Math.round(message.month);
    }
    if (message.enrollments !== 0) {
      obj.enrollments = Math.round(message.enrollments);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<EnrollmentTrend>, I>>(
    base?: I,
  ): EnrollmentTrend {
    return EnrollmentTrend.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<EnrollmentTrend>, I>>(
    object: I,
  ): EnrollmentTrend {
    const message = createBaseEnrollmentTrend();
    message.month = object.month ?? 0;
    message.enrollments = object.enrollments ?? 0;
    return message;
  },
};
// ...
export const InstructorCourseEnrollmentTrend: MessageFns<InstructorCourseEnrollmentTrend> =
  {
    encode(
      message: InstructorCourseEnrollmentTrend,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      for (const v of message.trend) {
        EnrollmentTrend.encode(v!, writer.uint32(10).fork()).join();
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): InstructorCourseEnrollmentTrend {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseInstructorCourseEnrollmentTrend();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.trend.push(EnrollmentTrend.decode(reader, reader.uint32()));
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },

    fromJSON(object: any): InstructorCourseEnrollmentTrend {
      return {
        trend: globalThis.Array.isArray(object?.trend)
          ? object.trend.map((e: any) => EnrollmentTrend.fromJSON(e))
          : [],
      };
    },

    toJSON(message: InstructorCourseEnrollmentTrend): unknown {
      const obj: any = {};
      if (message.trend?.length) {
        obj.trend = message.trend.map((e) => EnrollmentTrend.toJSON(e));
      }
      return obj;
    },

    create<I extends Exact<DeepPartial<InstructorCourseEnrollmentTrend>, I>>(
      base?: I,
    ): InstructorCourseEnrollmentTrend {
      return InstructorCourseEnrollmentTrend.fromPartial(base ?? ({} as any));
    },
    fromPartial<
      I extends Exact<DeepPartial<InstructorCourseEnrollmentTrend>, I>,
    >(object: I): InstructorCourseEnrollmentTrend {
      const message = createBaseInstructorCourseEnrollmentTrend();
      message.trend =
        object.trend?.map((e) => EnrollmentTrend.fromPartial(e)) || [];
      return message;
    },
  };
// ...
export const GetInstructorCourseEnrollmentTrendRequest: MessageFns<GetInstructorCourseEnrollmentTrendRequest> =
  {
    encode(
      message: GetInstructorCourseEnrollmentTrendRequest,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      if (message.courseId !== "") {
        writer.uint32(10).string(message.courseId);
      }
      if (message.instructorId !== "") {
        writer.uint32(34).string(message.instructorId);
      }
      if (message.from !== "") {
        writer.uint32(18).string(message.from);
      }
      if (message.to !== "") {
        writer.uint32(26).string(message.to);
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): GetInstructorCourseEnrollmentTrendRequest {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseGetInstructorCourseEnrollmentTrendRequest();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.courseId = reader.string();
            continue;
          }
          case 4: {
            if (tag !== 34) {
              break;
            }

            message.instructorId = reader.string();
            continue;
          }
          case 2: {
            if (tag !== 18) {
              break;
            }

            message.from = reader.string();
            continue;
          }
          case 3: {
            if (tag !== 26) {
              break;
            }

            message.to = reader.string();
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },

    fromJSON(object: any): GetInstructorCourseEnrollmentTrendRequest {
      return {
        courseId: isSet(object.courseId)
          ? globalThis.String(object.courseId)
          : isSet(object.course_id)
            ? globalThis.String(object.course_id)
            : "",
        instructorId: isSet(object.instructorId)
          ? globalThis.String(object.instructorId)
          : isSet(object.instructor_id)
            ? globalThis.String(object.instructor_id)
            : "",
        from: isSet(object.from) ? globalThis.String(object.from) : "",
        to: isSet(object.to) ? globalThis.String(object.to) : "",
      };
    },

    toJSON(message: GetInstructorCourseEnrollmentTrendRequest): unknown {
      const obj: any = {};
      if (message.courseId !== "") {
        obj.courseId = message.courseId;
      }
      if (message.instructorId !== "") {
        obj.instructorId = message.instructorId;
      }
      if (message.from !== "") {
        obj.from = message.from;
      }
      if (message.to !== "") {
        obj.to = message.to;
      }
      return obj;
    },

    create<
      I extends Exact<
        DeepPartial<GetInstructorCourseEnrollmentTrendRequest>,
        I
      >,
    >(base?: I): GetInstructorCourseEnrollmentTrendRequest {
      return GetInstructorCourseEnrollmentTrendRequest.fromPartial(
        base ?? ({} as any),
      );
    },
    fromPartial<
      I extends Exact<
        DeepPartial<GetInstructorCourseEnrollmentTrendRequest>,
        I
      >,
    >(object: I): GetInstructorCourseEnrollmentTrendRequest {
      const message = createBaseGetInstructorCourseEnrollmentTrendRequest();
      message.courseId = object.courseId ?? "";
      message.instructorId = object.instructorId ?? "";
      message.from = object.from ?? "";
      message.to = object.to ?? "";
      return message;
    },
  };
// ...
type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | boolean
  | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}